+++ packages/next/src/index.ts
@@ -1,22 +1,83 @@
-import { NextPageContext } from 'next'
+import {
+  GetServerSidePropsContext,
+  GetStaticPropsContext,
+  NextPageContext,
+} from 'next'
+import { ComponentType } from 'react'
 import { AppContext as NextAppContext } from 'next/app'
-import { AppContext } from '@fleur/fleur'
+import { action, actions, AppContext, operations } from '@fleur/fleur'
 import superjson from 'superjson'
 
-/** @deprecated Use `FleurishNextPageContext` instead */
-export interface PageContext extends NextPageContext {
+export interface FleurishNextPageContext extends NextPageContext {
   executeOperation: AppContext['executeOperation']
   getStore: AppContext['getStore']
   fleurContext: AppContext
 }
 
-export type FleurishNextPageContext = PageContext
+export type FleurishNextPage<P = {}, IP = P> = ComponentType<P & IP> & {
+  getInitialProps?(context: FleurishNextPageContext): IP | Promise<IP>
+}
 
 export interface FleurishNextAppContext extends NextAppContext {
   executeOperation: AppContext['executeOperation']
   getStore: AppContext['getStore']
   fleurContext: AppContext
 }
+
+export interface FleurishGetServerSidePropsContext
+  extends GetServerSidePropsContext {
+  executeOperation: AppContext['executeOperation']
+  getStore: AppContext['getStore']
+}
+export interface FleurishGetStaticPropsContext extends GetStaticPropsContext {
+  executeOperation: AppContext['executeOperation']
+  getStore: AppContext['getStore']
+}
+
+const __internalRehydrateAction = action<{ storeName: string; state: any }>(
+  '@fleur/next--internalRehydrate',
+)
+
+export const withSSPDistributer = (context: AppContext) => {
+  const { dispatch } = context
+
+  context.dispatch = (ident, payload) => {
+    if (ident !== __internalRehydrateAction) {
+      dispatch(ident, payload)
+      return
+    }
+
+    // Only dispatch to target Store
+    const { storeName, state } = payload
+
+    const listeners =
+      context
+        .getListenersOfStore(storeName)
+        ?.get(NextJsActions.rehydrateServerSideProps) ?? []
+
+    listeners.forEach((fn) => fn(state))
+  }
+
+  return context
+}
+
+export const NextJsOps = operations({
+  rehydrateServerSideProps: (
+    { dispatch },
+    { stores }: { stores: Record<string, any> },
+  ) => {
+    Object.keys(stores).forEach((storeName: string) => {
+      dispatch(__internalRehydrateAction, {
+        storeName,
+        state: stores[storeName],
+      })
+    })
+  },
+})
+
+export const NextJsActions = actions('@fleur/next/actions', {
+  rehydrateServerSideProps: action<any>(),
+})
 
 /** Add `executeOperation` and `getStore` method in NextAppContext */
 export const bindFleurContext = (
@@ -26,12 +87,12 @@ export const bindFleurContext = (
   // prettier-ignore
   ;(nextContext as FleurishNextAppContext).executeOperation
     = (nextContext.ctx as FleurishNextPageContext).executeOperation
-    = context.executeOperation.bind(context);
+    = context.executeOperation;
 
   // prettier-ignore
   ;(nextContext as FleurishNextAppContext).getStore
     = (nextContext.ctx as FleurishNextPageContext).getStore
-    = context.getStore.bind(context);
+    = context.getStore;
 
   // prettier-ignore
   ;(nextContext as FleurishNextAppContext).fleurContext